don-content-pipeline/pyproject.toml
[project]
name = "don-content-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "amsterdam-schema-tools>=7.3.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
